Restore flat_file test Remove lmdb_test, command_executor_test Move remove_all to common header
