+++ test/module/irohad/ametsuchi/ametsuchi_test_common.hpp
@@ -1,0 +1,49 @@
+/**
+ * Copyright Soramitsu Co., Ltd. 2017 All Rights Reserved.
+ * http://soramitsu.co.jp
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef IROHA_AMETSUCHI_TEST_COMMON_HPP
+#define IROHA_AMETSUCHI_TEST_COMMON_HPP
+
+#include <dirent.h>
+
+void remove_all(const std::string &dump_dir) {
+  if (!dump_dir.empty()) {
+    // Directory iterator:
+    struct dirent **namelist;
+    auto status = scandir(dump_dir.c_str(), &namelist, NULL, alphasort);
+    if (status < 0) {
+      // TODO: handle internal error
+    } else {
+      uint n = status;
+      uint i = 1;
+      while (++i < n) {
+        if (std::remove((dump_dir + "/" + namelist[i]->d_name).c_str())) {
+          perror("Error deleting file");
+        }
+      }
+      for (uint j = 0; j < n; ++j) {
+        free(namelist[j]);
+      }
+      free(namelist);
+    }
+    if (std::remove(dump_dir.c_str())) {
+      perror("Error deleting file");
+    }
+  }
+}
+
+#endif //IROHA_AMETSUCHI_TEST_COMMON_HPP
